UI homeAdmin & function Logout
