+++ src/main/webapp/WEB-INF/views/admin/home.html
@@ -11,6 +11,8 @@
             href="https://fonts.googleapis.com/css2?family=Roboto&amp;display=swap"
             rel="stylesheet"
     />
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/sweetalert2@11/dist/sweetalert2.min.css">
+    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
     <link th:href="@{/css/homeAdmin.css}" rel="stylesheet">
 </head>
 <body class="bg-[#e6f0f6] min-h-screen flex font-['Roboto']">
