+++ src/main/webapp/WEB-INF/views/authenticate/register.html
@@ -18,8 +18,8 @@
 <body class="bg-[#e6f0f6] min-h-screen flex items-center justify-center p-4">
 <div class="bg-[#d9e9f8] max-w-4xl w-full flex flex-col md:flex-row p-8 md:p-12">
     <div class="md:w-1/2 relative flex justify-center items-center border border-[#d9e9f8]">
-        <img src="https://storage.googleapis.com/a1aa/image/f625da54-0c5b-4945-ad3e-dfa294600676.jpg"
-             alt="Exterior view of a modern hospital building with clear blue sky"
+        <img th:src="@{/images/model_login.png}"
+             alt="Ảnh bệnh viện"
              class="object-cover w-full h-full" width="400" height="500"/>
     </div>
